Allow comma or semicolon separation in a list
This patch adds a new class, `AddressListParser`, with a single static
method `parse()`. It loops through each character of the value to
identify non-escaped, non-quoted delimiters, allowing either `,` or `;`
to be used.

`AbstractAddressList::fromString()` now uses the above method instead of
`str_getcsv()` to extract the list of addresses.

The patch also updates the test provided in #147 to test for the
existence of all addresses in the test string.