qa: remove laminas-config requirement in tests
The various storage adapters are really only looking for maps.
This patch adds a trait, `Laminas\Mail\Storage\ParamsNormalizerTrait`, which provides functionality for producing an array map from any of a `Traversable` object, a standard object, or an array (validating that only string keys are provided).
Each storage adapter constructor has been updated to use this and to then validate the values pulled from the array.
When done, there were 4 new `MixedAssignment` errors that needed to be added to the Psalm baseline.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>