Update Sendmail.php
Remove the \r\n translation for mail headers per the PHP mail() page. 

additional_headers (optional)
----------
String or array to be inserted at the end of the email header.

This is typically used to add extra headers (From, Cc, and Bcc). Multiple extra headers should be separated with a CRLF (\r\n). If outside data are used to compose this header, the data should be sanitized so that no unwanted headers could be injected.



Signed-off-by: cloudseeder <brooks@netgate.net>