Update testCanParseMultipartEmail
This change adds assertions for the content of the various parts of
message. The intention is to be able to check deeper into a multipart
message to confirm that it is as expected. As the underlying test data
contains a PDF file, it adds the smalot/pdfparser to simplify working
with PDF files.

Signed-off-by: Matthew Setter <matthew@matthewsetter.com>