Refactor instantiation of mime messages
From my understanding of the various specs, primarily RFC 1341, the
existing functionality for setting message headers was wrong. Given
that, I've made this change, which should bring it into line with the
spec, corrects issues with the existing tests, and sets one existing
test as skipped (for the time being).

Firstly, it removes the functionality that sets headers on the main
message from message sub-parts. Secondly, equally as importantly, it
sets the required headers, along with their defaults, based on the
message's composition. I won't get into detail here, as it's documented
in the code and in the updates to the documentation.

Finally, the change also cleans up some of the existing functionality as
well, setting return types, as they're supported in the package's
allowed PHP versions.

Signed-off-by: Matthew Setter <matthew@matthewsetter.com>