Add utility functions for retrieving message content and attachments
You can iterate over all of the message parts in the calling code and
extract the plain text and HTML body, as well as the attachments.
However, I thought that this would be something that people would do
regularly; so much so such that adding utility functions for doing so
might be a great productivity win for the package.

The documentation's been updated to reflect these additions so that
they're as visible as possible, and readily discoverable.

Signed-off-by: Matthew Setter <matthew@matthewsetter.com>